birthday-app/requirements.txt
flask==3.0.0
apscheduler==3.10.4
python-dotenv==1.0.0